+++ packages/indexer-nibi/src/heart-monitor.ts
@@ -1,9 +1,20 @@
 import { gqlEndptFromTmRpc } from "./gql"
+import { fundingRates, GqlInFundingRate, GqlOutFundingRate } from "./query/fundingRates"
+import {
+  GqlInLiquidation,
+  GqlOutLiquidations,
+  liquidations,
+} from "./query/liquidations"
 import {
   GqlInMarkPriceCandle,
   GqlOutMarkPriceCandle,
   markPriceCandles,
 } from "./query/markPriceCandles"
+import { GqlInMarkPrice, GqlOutMarkPrice, markPrices } from "./query/markPrices"
+import { GqlInTransfer, GqlOutTransfer, transfers } from "./query/transfer"
+import { GqlInOraclePrice, GqlOutOraclePrice, oraclePrices } from "./query/oraclePrices"
+import { GqlInPosition, GqlOutPosition, positions } from "./query/positions"
+import { GqlInUnbondings, GqlOutUnbondings, unbondings } from "./query/unbondings"
 
 /** IHeartMonitor is an interface for a Heart Monitor GraphQL API.
  * Each of its methods corresponds to a query function. */
@@ -11,6 +22,20 @@ export interface IHeartMonitor {
   readonly markPriceCandles: (
     args: GqlInMarkPriceCandle,
   ) => Promise<GqlOutMarkPriceCandle>
+
+  readonly markPrices: (args: GqlInMarkPrice) => Promise<GqlOutMarkPrice>
+
+  readonly liquidations: (args: GqlInLiquidation) => Promise<GqlOutLiquidations>
+
+  readonly fundingRates: (args: GqlInFundingRate) => Promise<GqlOutFundingRate>
+
+  readonly transfers: (args: GqlInTransfer) => Promise<GqlOutTransfer>
+
+  readonly oraclePrices: (args: GqlInOraclePrice) => Promise<GqlOutOraclePrice>
+
+  readonly positions: (args: GqlInPosition) => Promise<GqlOutPosition>
+
+  readonly unbondings: (args: GqlInUnbondings) => Promise<GqlOutUnbondings>
 
   /*
   readonly useQueryBlockMarkPrices: (args: {
@@ -67,6 +92,27 @@ export class HeartMonitor implements IHeartMonitor {
   markPriceCandles = async (
     args: GqlInMarkPriceCandle,
   ): Promise<GqlOutMarkPriceCandle> => markPriceCandles(args, this.gqlEndpt)
+
+  markPrices = async (args: GqlInMarkPrice): Promise<GqlOutMarkPrice> =>
+    markPrices(args, this.gqlEndpt)
+
+  liquidations = async (args: GqlInLiquidation): Promise<GqlOutLiquidations> =>
+    liquidations(args, this.gqlEndpt)
+
+  fundingRates = async (args: GqlInFundingRate): Promise<GqlOutFundingRate> =>
+    fundingRates(args, this.gqlEndpt)
+
+  transfers = async (args: GqlInTransfer): Promise<GqlOutTransfer> =>
+    transfers(args, this.gqlEndpt)
+
+  oraclePrices = async (args: GqlInOraclePrice): Promise<GqlOutOraclePrice> =>
+    oraclePrices(args, this.gqlEndpt)
+
+  positions = async (args: GqlInPosition): Promise<GqlOutPosition> =>
+    positions(args, this.gqlEndpt)
+
+  unbondings = async (args: GqlInUnbondings): Promise<GqlOutUnbondings> =>
+    unbondings(args, this.gqlEndpt)
 
   /*
   // ------------------------------------------------------------